+++ src/citra_qt/debugger/graphics/graphics.cpp
@@ -3,7 +3,7 @@
 // Refer to the license.txt file included.
 
 #include <QListView>
-#include "citra_qt/debugger/graphics.h"
+#include "citra_qt/debugger/graphics/graphics.h"
 #include "citra_qt/util/util.h"
 
 extern GraphicsDebugger g_debugger;
