+++ src/citra_qt/debugger/graphics/graphics.h
