+++ src/citra_qt/debugger/graphics/graphics_breakpoint_observer.h
