+++ src/citra_qt/debugger/graphics/graphics_vertex_shader.h
@@ -6,7 +6,7 @@
 
 #include <QAbstractTableModel>
 #include <QTreeView>
-#include "citra_qt/debugger/graphics_breakpoint_observer.h"
+#include "citra_qt/debugger/graphics/graphics_breakpoint_observer.h"
 #include "nihstro/parser_shbin.h"
 #include "video_core/shader/shader.h"
