+++ src/citra_qt/main.cpp
@@ -17,12 +17,12 @@
 #include "citra_qt/configure_dialog.h"
 #include "citra_qt/debugger/callstack.h"
 #include "citra_qt/debugger/disassembler.h"
-#include "citra_qt/debugger/graphics.h"
-#include "citra_qt/debugger/graphics_breakpoints.h"
-#include "citra_qt/debugger/graphics_cmdlists.h"
-#include "citra_qt/debugger/graphics_surface.h"
-#include "citra_qt/debugger/graphics_tracing.h"
-#include "citra_qt/debugger/graphics_vertex_shader.h"
+#include "citra_qt/debugger/graphics/graphics.h"
+#include "citra_qt/debugger/graphics/graphics_breakpoints.h"
+#include "citra_qt/debugger/graphics/graphics_cmdlists.h"
+#include "citra_qt/debugger/graphics/graphics_surface.h"
+#include "citra_qt/debugger/graphics/graphics_tracing.h"
+#include "citra_qt/debugger/graphics/graphics_vertex_shader.h"
 #include "citra_qt/debugger/profiler.h"
 #include "citra_qt/debugger/ramview.h"
 #include "citra_qt/debugger/registers.h"
